models, routes, relations updated

+++ routes/user.routes.js
@@ -3,46 +3,83 @@ const User = require("../models/user.model")
 // Recommendation model added to update totalSaved
 const Recommendation = require("../models/recommendation.model")
 
+const { verifyToken } = require("../middlewares/auth.middlewares")
 
-// TEMPORARY !!!
-// | POST        | `/api/users             | TESTING ONLY -- Creates a user                     |
-router.post("/", async (req, res, next) => {
+
+// // TEMPORARY !!!
+// // | POST        | `/api/users             | TESTING ONLY -- Creates a user                     |
+// router.post("/", async (req, res, next) => {
+//   try {
+//     const newUser = await User.create({
+//       username: req.body.username,
+//       email: req.body.email,
+//       socialLink: req.body.socialLink,
+//       savedRecs: [] // empty for new users
+//     })
+//     res.status(201).json(newUser)
+//     // console.log(newUser)
+//   } catch (error) {
+//     next(error)
+//   }
+// })
+
+
+// | GET         | `/api/users/:userId`    | Read a specific user's profile + RECS -- users only |
+router.get("/:userId", verifyToken, async (req, res, next) => {
   try {
-    const newUser = await User.create({
-      username: req.body.username,
-      email: req.body.email,
-      socialLinks: req.body.socialLinks,
-      savedRecs: [] // empty for new users
-    })
-    res.status(201).json(newUser)
-    // console.log(newUser)
+    const specificUser = await User.findById(req.params.userId)
+      .populate("createdRecs")
+    res.status(200).json(specificUser)
   } catch (error) {
     next(error)
   }
 })
 
 
-// | GET         | `/api/users/:userId`    | Read a specific user's profile                     |
-router.get("/:userId", async (req, res, next) => {
+// // this route could be unnecessary now that we have createdRecs property
+// // | GET | `/api/users/:userId/created-recommendations` | Get all recommendations created by a user -- users only |
+// router.get("/:userId/created-recommendations", verifyToken, async (req, res, next) => {
+//   try {
+//     const createdRecs = await Recommendation.find({ creator: req.params.userId })
+//     .populate("content") // bring content data
+//     res.status(200).json(createdRecs)
+//   } catch (error) {
+//     next(error)
+//   }
+// })
+
+
+// | GET | `/api/users/:userId/saved-recommendations` | Get all saved recommendations of a user (private) |
+router.get("/:userId/saved-recommendations", verifyToken, async (req, res, next) => {
   try {
-    const specificUser = await User.findById(req.params.userId)
-      .populate("savedRecs") // populates recommendations data when fetching the user
-    res.status(200).json(specificUser);
+    if (req.payload._id !== req.params.userId) {
+      return res.status(403).json({ message: "You cannot access someone else's saved recommendations." });
+    }
+
+    const user = await User.findById(req.params.userId).populate("savedRecs");
+    res.status(200).json(user.savedRecs)
   } catch (error) {
-    next(error);
+    next(error)
   }
 })
 
 
-// | PUT         | `/api/users/:userId`    | Update a specific user's profile                   |
-router.put("/:userId", async (req, res, next) => {
+
+
+// | PUT         | `/api/users/:userId`    | Update a specific user's profile -- if it's owned                  |
+router.put("/:userId", verifyToken, async (req, res, next) => {
   try {
+
+    if (req.payload._id !== req.params.userId) {
+      return res.status(403).json({ message: "you cannot change someone -- else's profile <3" });
+    }
+
     const updatedUser = await User.findByIdAndUpdate(
       req.params.userId,
       {
         username: req.body.username,
         email: req.body.email,
-        socialLinks: req.body.socialLinks
+        socialLink: req.body.socialLink
         // savedRecs: [] // removed cause it won't be updated here - is this ok ?
         // add more fields if you change the model !!!
       },
@@ -90,7 +127,7 @@ router.put("/:userId/save/:recommendationId", async (req, res, next) => {
 router.put("/:userId/unsave/:recommendationId", async (req, res, next) => {
   try {
 
-    const user = await User.findById(req.params.userId);
+    const user = await User.findById(req.params.userId)
 
     // remove recommendation from the user's saved list
     // use the recId params to return all array elements but that one
@@ -114,9 +151,14 @@ router.put("/:userId/unsave/:recommendationId", async (req, res, next) => {
 })
 
 
-// | DELETE      | `/api/users/:userId`    | Delete a specific user's account                   |
-router.delete("/:userId", async (req, res, next) => {
+// | DELETE      | `/api/users/:userId`    | Delete a specific user's account -- owned                   |
+router.delete("/:userId", verifyToken, async (req, res, next) => {
   try {
+
+    if (req.params.userId.toString() !== req.payload._id) {
+      return res.status(403).json({ message: "you cannot delete this, it's not yours" })
+    }
+
     await User.findByIdAndDelete(req.params.userId)
     res.sendStatus(204)
   } catch (error) {
